Implemented basic tile grid generation
MapGen will eventually be used to configure the map generation parameters.  At this moment it only allows configuration of map size.

MapTile Prefab is set on the MapManager.cs component

Prefabs are spawned such that they don't overlap, and have no space between them.

MapTile.cs component is used to store information about each tile.  It should be included on any prefab that is used as a tile.

MapGenerator.prefab contains the components which call the map generation functions.